requirements.txt
asgiref==3.4.1
Django==4.0.4
django-environ==0.8.1
django-extensions==3.1.3
django-filter==21.1
djangorestframework==3.12.4
pathlib==1.0.1
psycopg2-binary==2.9.1
pydantic==1.9.1
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing_extensions==4.2.0
